move autofac to own folder
