+++ Src/MoneyManager.Core/AutoFac/AutofacMvxIocProviderResources.resx
